move all functions related to formatting evidence/alerts text to their own class
